Fixed missing handle index
With pinsplash's help, fixed the "missing index" issue and am now passing around handles through EntFire.

Looks like the deletion is failing though. My guess is that once I destroy the handle, I can no longer use it as an index into a table, so I have to clean that up first and then destroy it.