Replaced parent
Replaced the parent "info_target" with a invisible "func_brush" instead (which in tf2 is not a preserved entity and does not have weird round start behavior). I still see it visually teleport into place at the beginning of the round though, which is undesirable, but I can't rotate the brush statically.